Splitted Chart.py
cleanups
